+++ backend/routes/api/spots.js
@@ -16,17 +16,31 @@ router.get('/current', requireAuth, async (req, res) => {
 
 // #22 ; /:spotId/images/:imageId ; DELETE
 router.delete('/:spotId/images/:imageId', requireAuth, async (req, res, next) => {
+    const currSpotImage = await SpotImage.findAll({where: {
+        id: req.params.imageId,
+        spotId: req.params.spotId
+    }});
+    if (currSpotImage) {
+        const err = new Error("Spot Image couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
     const imageToDelete = await SpotImage.destroy({where: {
         id: req.params.imageId,
         spotId: req.params.spotId
     }});
     res.status(200);
     res.json({message: 'Successfully deleted'});
-    //err: 404, 'Review Image couldn't be found'
 });
 
 // #8 ; /:spotId/images ; POST
 router.post('/:spotId/images', requireAuth, async (req, res, next) => {
+    const currSpot = await Spot.findByPk(req.params.spotId);
+    if (!currSpot) {
+        const err = new Error("Spot couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
     const {url, preview} = req.body;
     const spotId = req.params;
     const newSpotImage = await SpotImage.create({
@@ -36,21 +50,40 @@ router.post('/:spotId/images', requireAuth, async (req, res, next) => {
     });
     res.status(200);
     res.json(newSpotImage);
-    //err: 404, 'Spot couldn't be found'
 });
 
 // #12 ; /:spotId/reviews ; GET
 router.get('/:spotId/reviews', async (req, res, next) => {
+    const currSpot = await Spot.findByPk(req.params.spotId);
+    if (!currSpot) {
+        const err = new Error("Review Image couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
     const reviewsBySpot = await Review.findAll({where: {spotId: req.params.spotId}});
     res.status(200);
     res.json(reviewsBySpot); //add User, ReviewImages
-    //err: 404, 'Spot couldn't be found'
 });
 
 // #13 ; /:spotId/reviews ; POST
 router.post('/:spotId/reviews', requireAuth, validateData, async (req, res, next) => {
+    const currSpot = await Spot.findByPk(req.params.spotId);
+    if (!currSpot) {
+        const err = new Error("Review Image couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
     const {user} = req;
     const userId = user.id;
+    const existingReview = await Review.findAll({where: {
+        userId: userId,
+        spotId: req.params.spotId
+    }});
+    if (existingReview) {
+        const err = new Error("User already has a review for this spot");
+        err.status = 500;
+        return next(err);
+    }
     const {review, stars} = req.body;
     const newReview = await Review.create({
         userId: userId,
@@ -60,8 +93,6 @@ router.post('/:spotId/reviews', requireAuth, validateData, async (req, res, next
     });
     res.status(201);
     res.json(newReview);
-    //err: 404, 'Spot couldn't be found'
-    //err: 500, 'User already has a review for this spot'
 });
 
 // #18 ; /:spotId/bookings ; GET
@@ -73,20 +104,25 @@ router.get('/:spotId/bookings', requireAuth, async (req, res) => {
 
 // #19 ; /:spotId/bookings ; POST
 router.post('/:spotId/bookings', requireAuth, async (req, res, next) => {
+    const currSpot = await Spot.findByPk(req.params.spotId);
+    if (!currSpot) {
+        const err = new Error("Review Image couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
     const {user} = req;
     const userId = user.id;
-    const {startDate, endDate} = req.body;
-    const spotId = req.params;
-    const newBooking = await Booking.create({
-        spotId: spotId,
-        userId: userId,
-        startDate: startDate,
-        endDate: endDate
-    });
-    res.status(200);
-    res.json(newBooking);
-    //err: 404, 'Spot couldn't be found'
-    //err: 403,
+    const {requestedStartDate, requestedEndDate} = req.body;
+    const existingBooking = await Booking.findAll({where: {
+        spotId: req.params.spotId,
+        startDate: //before requestedEndDate
+        endDate: //after requestStartDate
+    }});
+    if (existingBooking) {
+        const err = new Error("Sorry, this spot is already booked for the specified dates");
+        err.status = 403;
+        return next(err);
+    }
     // {
     //     "message": "Sorry, this spot is already booked for the specified dates",
     //     "errors": {
@@ -94,20 +130,37 @@ router.post('/:spotId/bookings', requireAuth, async (req, res, next) => {
     //       "endDate": "End date conflicts with an existing booking"
     //     }
     // }
+    const newBooking = await Booking.create({
+        spotId: req.params.spotId,
+        userId: userId,
+        startDate: requestedStartDate,
+        endDate: requestedEndDate
+    });
+    res.status(200);
+    res.json(newBooking);
 });
 
 // #6 ; /:spotId/ ; GET
 router.get('/:spotId', async (req, res, next) => {
     const spotToGet = await Spot.findByPk(req.params.spotId);
+    if (!spotToGet) {
+        const err = new Error("Spot couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
     res.status(200);
     res.json(spotToGet); //add numReviews, avgStarRating, SpotImages, Owner
-    //err: 404, 'Spot couldn't be found'
 });
 
 // #9 ; /:spotId/ ; PUT
 router.put('/:spotId', requireAuth, async (req, res, next) => {
-    const {address, city, state, country, lat, lng, name, description, price} = req.body;
     const spotToUpdate = await Spot.findByPk(req.params.spotId);
+    if (!spotToUpdate) {
+        const err = new Error("Spot couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
+    const {address, city, state, country, lat, lng, name, description, price} = req.body;
     spotToUpdate.address = address;
     spotToUpdate.city = city;
     spotToUpdate.state = state;
@@ -119,15 +172,19 @@ router.put('/:spotId', requireAuth, async (req, res, next) => {
     spotToUpdate.price = price;
     res.status(200);
     res.json(spotToUpdate);
-    //err: 404, 'Spot couldn't be found'
 });
 
 // #10 ; /:spotId/ ; DELETE
 router.delete('/:spotId', requireAuth, async (req, res, next) => {
+    const currSpot = await Spot.findByPk(req.params.spotId);
+    if (!currSpot) {
+        const err = new Error("Spot couldn't be found");
+        err.status = 404;
+        return next(err);
+    }
     const spotToDelete = await Spot.destroy({where: {id: req.params.spotId}});
     res.status(200);
     res.json({message: 'Successfully deleted'});
-    //err: 404, 'Spot couldn't be found'
 });
 
 // #4 ; / ; GET
